backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.10.0
google-generativeai==0.8.6
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
python-dateutil==2.8.2
pytz==2024.1
pytest==7.4.3
pytest-flask==1.3.0
APScheduler==3.10.4
firebase-admin==6.5.0
markdownify==1.2.2
Pillow==10.4.0
